refactor: Extract logic to create ElasticsearchResult from StatementService into own Service.
